static_cast

+++ src/problems/RadMarshak/test_radiation_marshak.cpp
@@ -116,7 +116,7 @@ AMRSimulation<SuOlsonProblem>::setCustomBoundaryConditions(const amrex::IntVect 
 							   amrex::GeometryData const & /*geom*/, const amrex::Real /*time*/, const amrex::BCRec *bcr,
 							   int /*bcomp*/, int /*orig_comp*/)
 {
-	if ((bcr->lo(0) != quokka::BCType::ext_dir) && (bcr->hi(0) != quokka::BCType::ext_dir)) {
+	if ((bcr->lo(0) != static_cast<int>(quokka::BCType::ext_dir)) && (bcr->hi(0) != static_cast<int>(quokka::BCType::ext_dir))) {
 		return;
 	}
